fix(preflight): code quality improvements across gpu_memory and preflight modules
- Add negative device_index validation in measure.py
- Remove redundant pass statement in query.py placeholder class
- Fix missing space in error message in query.py
- Extract shared _exit_code_for_status helper in preflight/__init__.py
- Add wildcard fallback to _generate_message in check.py
- Add exit_code field to mini_run JSON output for both success and failure
- Remove redundant int() wrapper in test_full_check.py
- Add per-file-ignore for E402 in preflight/__init__.py (circular import pattern)

+++ tests/unit/preflight/test_full_check.py
@@ -501,7 +501,7 @@ class TestFullPreflightResultExitCodes:
         )
         mock_subprocess.return_value = {
             "success": True,
-            "peak_memory_bytes": int(9 * 1024**3),  # 9 GB of 10 GB = low headroom
+            "peak_memory_bytes": 9 * 1024**3,  # 9 GB of 10 GB = low headroom
             "runtime_seconds": 5.0,
         }
